+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- example/5/draggable.js fix for FireFox to prevent dropping a draggable trying to load a URL
+
 ## [v4.1.1] 
 ### Removed
 - ppc64le and i386 builds of github, because ppc64le keep failing mysteriously, and we don't need them anyhow.